feat(market-data): preserve fractional volume

+++ src/invest/adapters/fixtures_json.py
@@ -33,7 +33,7 @@ class _BarPayload(BaseModel):
     high: Decimal = Field(gt=0)
     low: Decimal = Field(gt=0)
     close: Decimal = Field(gt=0)
-    volume: int = Field(ge=0)
+    volume: Decimal = Field(ge=0)
 
     @model_validator(mode="after")
     def validate_price_relationships(self) -> "_BarPayload":
